# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmah - Sports T-Shirt Customization API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Team {
  id: number;
  /** Team name in Arabic */
  name: string;
  /** Team name in English */
  nameEn: string;
  /** League name */
  league: string;
  country: string;
  /** Hex color */
  primaryColor: string;
  /** Hex color */
  secondaryColor: string;
  /** Available jersey colors as hex codes */
  availableColors: string[];
  /** Available sizes (XS, S, M, L, XL, XXL) */
  availableSizes: string[];
  basePrice: number;
  logoUrl?: string;
  /** Number of orders for this team */
  orderCount: number;
}

export type CreateOrderBodySize =
  (typeof CreateOrderBodySize)[keyof typeof CreateOrderBodySize];

export const CreateOrderBodySize = {
  XS: "XS",
  S: "S",
  M: "M",
  L: "L",
  XL: "XL",
  XXL: "XXL",
} as const;

export interface CreateOrderBody {
  teamId: number;
  /** Name to print on jersey */
  customerName: string;
  /** Number to print on jersey */
  jerseyNumber: string;
  size: CreateOrderBodySize;
  /** Selected jersey color (hex) */
  color: string;
  /** @minimum 1 */
  quantity: number;
  customerPhone: string;
  customerCity: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
} as const;

export interface Order {
  id: number;
  teamId: number;
  teamName: string;
  customerName: string;
  jerseyNumber: string;
  size: string;
  color: string;
  quantity: number;
  totalPrice: number;
  customerPhone: string;
  customerCity: string;
  status: OrderStatus;
  createdAt: string;
}

export interface OrderStats {
  totalOrders: number;
  totalRevenue: number;
  topTeam: string;
  popularSize: string;
  popularColor: string;
}