# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Basmah - Sports T-Shirt Customization API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Team name in Arabic
        nameEn:
          type: string
          description: Team name in English
        league:
          type: string
          description: League name
        country:
          type: string
        primaryColor:
          type: string
          description: Hex color
        secondaryColor:
          type: string
          description: Hex color
        availableColors:
          type: array
          items:
            type: string
          description: Available jersey colors as hex codes
        availableSizes:
          type: array
          items:
            type: string
          description: Available sizes (XS, S, M, L, XL, XXL)
        basePrice:
          type: number
        logoUrl:
          type: string
        orderCount:
          type: integer
          description: Number of orders for this team
      required:
        - id
        - name
        - nameEn
        - league
        - country
        - primaryColor
        - secondaryColor
        - availableColors
        - availableSizes
        - basePrice
        - orderCount
...