# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basmah - Sports T-Shirt Customization API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  /** Team name in Arabic */
  name: string;
  /** Team name in English */
  nameEn: string;
  /** League name */
  league: string;
  country: string;
  /** Hex color */
  primaryColor: string;
  /** Hex color */
  secondaryColor: string;
  /** Available jersey colors as hex codes */
  availableColors: string[];
  /** Available sizes (XS, S, M, L, XL, XXL) */
  availableSizes: string[];
  basePrice: number;
  logoUrl?: string;
  /** Number of orders for this team */
  orderCount: number;
}